requirements.txt
aiogram>=3.7.0
APScheduler>=3.10.0
pytz>=2024.1
python-dotenv>=1.0.0
